[TASK] Removal of AnnotationService and everything which depends on it
As a relict, we still had the AnnotationService in place; although we wanted
to not use it anymore and replace it with standard FLOW3 functionality.

This method removes everything related to the AnnotationService. This might break
some functionality like the (not-fully-working) inline editing.

We need to re-implement this functionality then, based on the standard FLOW3
ReflectionService.

Change-Id: I98229b3bba2e3efa6327e8c65a6f995998b3f146